% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData
Provides a General dataset of daily and Cumulative cases and deaths of covid-19}
\usage{
getData()
}
\value{
a DataFrame grouped by Country and Date has the number of deaths and cases and daily deaths and daily cases
}
\description{
getData
Provides a General dataset of daily and Cumulative cases and deaths of covid-19
}
\examples{
df <- getData()
}
