% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOnWorldMap.R
\name{plotOnWorldMap}
\alias{plotOnWorldMap}
\title{plotOnWorldMap}
\usage{
plotOnWorldMap(date, type)
}
\arguments{
\item{date, }{a string on format 'year-month-day'}

\item{type, }{a string which can be 'Cases', 'Deaths', 'DailyCases', 'DailyDeaths'}
}
\description{
Draw a World map which each countrie's color(fill) showing number of selected type(Daily Cases, Daily Deaths, All Deaths, All Cases)
}
