% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{plotTimeSeries}
\usage{
plotTimeSeries(startDate, endDate, country, type)
}
\arguments{
\item{startDate, }{a string on format 'year-month-day', shows start date}

\item{endDate, }{a string on format 'year-month-day', shows end date. endDate must be greater than startDate}

\item{country, }{a string of selected country}

\item{type, }{a string which can be 'Confirmed', 'Deaths', 'DailyConfirmed', 'DailyDeaths'}
}
\description{
Draw a time series line plot based on number of type(Daily Cases, Daily Deaths, All Deaths, All Cases)
}
\examples{
plotTimeSeries('2020-9-21', '2020-10-21', 'US', 'DailyConfirmed')
}
