% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(saveData = T)
}
\value{
a DataFrame grouped by Country and Date has the number of deaths and cases and daily deaths and daily cases
}
\description{
Provides a General dataset of daily and Cumulative cases and deaths of covid-19
}
